Move methods to AbstractRule class #28
